rename `litdata`
