+++ MANIFEST.in
@@ -5,10 +5,10 @@ recursive-exclude __pycache__  *.py[cod] *.orig
 
 # Include the README and CHANGELOG
 include *.md
-recursive-include lightning_data *.md
+recursive-include litdata *.md
 
 # Include the code
-recursive-include lightning_data *.py
+recursive-include litdata *.py
 
 # Include the license file
 include LICENSE
