+++ litdata/__init__.py
@@ -1,0 +1,20 @@
+from lightning_utilities.core.imports import RequirementCache
+
+from litdata.__about__ import *  # noqa: F403
+from litdata.processing.functions import map, optimize, walk
+from litdata.streaming.combined import CombinedStreamingDataset
+from litdata.streaming.dataloader import StreamingDataLoader
+from litdata.streaming.dataset import StreamingDataset
+
+__all__ = [
+    "StreamingDataset",
+    "CombinedStreamingDataset",
+    "StreamingDataLoader",
+    "map",
+    "optimize",
+    "walk",
+]
+if RequirementCache("lightning_sdk"):
+    from lightning_sdk import Machine  # noqa: F401
+
+    __all__ + ["Machine"]
