+++ litdata/processing/readers.py
