+++ litdata/streaming/compression.py
