+++ pyproject.toml
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 [metadata]
-name = "lightning_data"
+name = "litdata"
 author = "Lightning-AI et al."
 url = "https://github.com/Lightning-AI/lit-data"
 
@@ -87,7 +87,7 @@ lint.ignore-init-module-imports = true
     "S501",  # Probable use of `requests` call with `verify=False` disabling SSL certificate checks
     "S108",  # Probable insecure usage of temporary file or directory: "/tmp/data/MNIST"
 ]
-"lightning_data/**" = [
+"litdata/**" = [
     "S101",  # todo: Use of `assert` detected
     "S105", "S106", "S107",  # todo: Possible hardcoded password: ...
     "S113",  # todo: Probable use of requests call without timeout
@@ -131,7 +131,7 @@ max-complexity = 10
 
 [tool.mypy]
 files = [
-    "lightning_data",
+    "litdata",
 ]
 # This section is for folders with "-" as they are not valid python modules
 exclude = [
