+++ tests/processing/test_data_processor.py
@@ -12,9 +12,9 @@ import torch
 from lightning import seed_everything
 from lightning_utilities.core.imports import RequirementCache
 
-from lightning_data.processing import data_processor as data_processor_module
-from lightning_data.processing import functions
-from lightning_data.processing.data_processor import (
+from litdata.processing import data_processor as data_processor_module
+from litdata.processing import functions
+from litdata.processing.data_processor import (
     DataChunkRecipe,
     DataProcessor,
     DataTransformRecipe,
@@ -27,9 +27,9 @@ from lightning_data.processing.data_processor import (
     _wait_for_disk_usage_higher_than_threshold,
     _wait_for_file_to_exist,
 )
-from lightning_data.processing.functions import LambdaDataTransformRecipe, map, optimize
-from lightning_data.streaming import resolver
-from lightning_data.streaming.cache import Cache, Dir
+from litdata.processing.functions import LambdaDataTransformRecipe, map, optimize
+from litdata.streaming import resolver
+from litdata.streaming.cache import Cache, Dir
 
 _PIL_AVAILABLE = RequirementCache("PIL")
 
@@ -164,7 +164,7 @@ def test_remove_target(tmpdir):
 
 
 @pytest.mark.skipif(condition=sys.platform == "win32", reason="Not supported on windows")
-@mock.patch("lightning_data.processing.data_processor._wait_for_disk_usage_higher_than_threshold")
+@mock.patch("litdata.processing.data_processor._wait_for_disk_usage_higher_than_threshold")
 def test_download_data_target(wait_for_disk_usage_higher_than_threshold_mock, tmpdir):
     input_dir = os.path.join(tmpdir, "input_dir")
     os.makedirs(input_dir, exist_ok=True)
@@ -203,7 +203,7 @@ def test_download_data_target(wait_for_disk_usage_higher_than_threshold_mock, tm
 
 def test_wait_for_disk_usage_higher_than_threshold():
     disk_usage_mock = mock.Mock(side_effect=[mock.Mock(free=10e9), mock.Mock(free=10e9), mock.Mock(free=10e11)])
-    with mock.patch("lightning_data.processing.data_processor.shutil.disk_usage", disk_usage_mock):
+    with mock.patch("litdata.processing.data_processor.shutil.disk_usage", disk_usage_mock):
         _wait_for_disk_usage_higher_than_threshold("/", 10, sleep_time=0)
     assert disk_usage_mock.call_count == 3
