+++ tests/streaming/test_cache.py
@@ -25,12 +25,12 @@ from lightning_utilities.core.imports import RequirementCache
 from lightning_utilities.test.warning import no_warning_call
 from torch.utils.data import Dataset
 
-from lightning_data.streaming import Cache
-from lightning_data.streaming.dataloader import CacheDataLoader
-from lightning_data.streaming.dataset import StreamingDataset
-from lightning_data.streaming.item_loader import TokensLoader
-from lightning_data.streaming.serializers import Serializer
-from lightning_data.utilities.env import _DistributedEnv
+from litdata.streaming import Cache
+from litdata.streaming.dataloader import CacheDataLoader
+from litdata.streaming.dataset import StreamingDataset
+from litdata.streaming.item_loader import TokensLoader
+from litdata.streaming.serializers import Serializer
+from litdata.utilities.env import _DistributedEnv
 
 _PIL_AVAILABLE = RequirementCache("PIL")
 _TORCH_VISION_AVAILABLE = RequirementCache("torchvision")
