+++ tests/streaming/test_serializer.py
@@ -22,7 +22,7 @@ import torch
 from lightning import seed_everything
 from lightning_utilities.core.imports import RequirementCache
 
-from lightning_data.streaming.serializers import (
+from litdata.streaming.serializers import (
     _AV_AVAILABLE,
     _NUMPY_DTYPES_MAPPING,
     _SERIALIZERS,
