+++ tests/utilities/test_packing.py
@@ -1,6 +1,6 @@
 import pytest
 
-from lightning_data.utilities.packing import _pack_greedily
+from litdata.utilities.packing import _pack_greedily
 
 
 def test_pack_greedily():
